+++ src/component/multibus/multibus.cpp
@@ -7,6 +7,7 @@
     multibus.cpp: The Intel Multibus bus implementation.
 */
 
+#include <base/emulation.hpp>
 #include <component/multibus/multibus.hpp>
 
 namespace Motion
@@ -21,12 +22,26 @@ namespace Motion
 
         AddrSpace::AddMapping(mappingMultibus);
 
+        // map ip2 segment 5 (multibus IO)
+
         AddrSpaceMapping mappingIo = AddrSpaceMapping();
         mappingIo.startAddr = MULTIBUS_IO_START;
         mappingIo.endAddr = MULTIBUS_IO_END;
 
         AddrSpace::AddMapping(mappingIo);
         
-        // map ip2 segment 5 (multibus IO)
+        if (!cpu)
+            cpu = Emulation::GetMachine().FindComponentByType<ComponentCPU>();
+    }
+
+    void Multibus::FireMultibusIRQ(int32_t number)
+    {
+        if (number > MULTIBUS_NUM_IRQ)
+        {
+            Logger::Log(MULTIBUS_LOG_PREFIX, std::format("Tried to fire invalid IRQ #{}", number).c_str(), LogChannels::Warning);
+            return;
+        }
+
+        cpu->SetIRQLine(number);
     }
 };